fix(api): make bodyfat router import stable
